requirements.txt
pysnyk==0.9.19
yaspin==3.0.2
